fix: updated organization, small bugs
